% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hilmo_import_csv.R
\name{Hilmo_import_csv}
\alias{Hilmo_import_csv}
\title{A Hilmo import .csv function}
\usage{
Hilmo_import_csv(filename)
}
\arguments{
\item{filename}{the name of your file you wish to import}
}
\description{
This function allows you to 
  - import csv files, leaving only columns: SUKUP,IKA, TMPKOODI, KOODI1, lahtopvm, tnro, PALTU, KOKU 
  - Adds ID column from lahtopaiva and tnro
  - Adds "year" column
}
\examples{
Hilmo_import_csv()
}
\keyword{csv}
\keyword{hilmo}
\keyword{import}
