% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hilmo_tidy.R
\name{Hilmo_tidy}
\alias{Hilmo_tidy}
\title{A Hilmo tidy function}
\usage{
Hilmo_tidy(dataname)
}
\arguments{
\item{dataname}{the name of your data you are working with}
}
\description{
This function allows you to change the data from long to wide format
}
\details{
Every row represents one patient
}
\examples{
Hilmo_tidy()
}
\keyword{hilmo}
\keyword{tidy}
