% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hilmo_tidy_hoitojakso.R
\name{Hilmo_tidy_hoitojakso}
\alias{Hilmo_tidy_hoitojakso}
\title{A Hilmo hoitojakso tidy function}
\usage{
Hilmo_tidy_hoitojakso(dataname)
}
\arguments{
\item{dataname}{the name of your data you are working with}
}
\description{
This function allows you to change the data from long to wide format by adding all diagnose codes into one row.
}
\details{
Every row represents one hoitojakso
}
\examples{
Hilmo_tidy_hoitojakso()

}
\keyword{hilmo}
\keyword{hoitojakso}
\keyword{tidy}
